LayoutTests/editing/pasteboard/copy-content-with-user-select-none-expected.txt

This tests copying excludes content with user-select: none.
To manually test, copy "hello world foo bar" below then paste.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString().includes("hello") is true
PASS getSelection().toString().includes("world") is false
PASS getSelection().toString().includes("foo") is false
PASS getSelection().toString().includes("bar") is true
PASS event.clipboardData.getData("text/plain").includes("hello") is true
PASS event.clipboardData.getData("text/plain").includes("world") is false
PASS event.clipboardData.getData("text/plain").includes("foo") is false
PASS event.clipboardData.getData("text/plain").includes("bar") is true
PASS event.clipboardData.getData("text/html").includes("hello") is true
PASS event.clipboardData.getData("text/html").includes("world") is false
PASS event.clipboardData.getData("text/html").includes("foo") is false
PASS event.clipboardData.getData("text/html").includes("bar") is true
PASS successfullyParsed is true

TEST COMPLETE
hello world foo bar
hello bar
